adding the retrieval and useage of the swiss-prot curated uniprotkb id filtering, cleaning up logging, trapping the creation of a graph edge if the nodes werent normalized

+++ GOA/src/loadGOA.py
@@ -4,10 +4,9 @@ import argparse
 import enum
 import pandas as pd
 import gzip
-from io import TextIOBase, TextIOWrapper
+import logging
+from io import TextIOWrapper
 from csv import reader
-from operator import itemgetter
-from zipfile import ZipFile
 from Common.utils import LoggingUtil, GetData, NodeNormUtils
 from pathlib import Path
 
@@ -44,26 +43,29 @@ class DATACOLS(enum.IntEnum):
 # Desc: Class that loads the UniProtKB GOA data and creates KGX files for importing into a Neo4j graph.
 ##############
 class GOALoader:
-    def load(self, data_file_path: str, data_file_name: str, out_name: str) -> bool:
+    def load(self, data_file_path, ftp_dir_path: str, data_file_name: str, out_name: str) -> bool:
         """
-        loads/parses goa data file from ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/HUMAN/
+        loads/parses goa data file from ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/<ftp_dir_path>
 
         :param data_file_path: root directory of output data files
+        :param ftp_dir_path: the path in the ftp file sub directory
         :param data_file_name: the name of the goa input data file
         :param out_name: the output name prefix of the KGX files
         :return: True
         """
+        logger.info(f'GOALoader - Start of file processing.')
 
         # init the return flag
         ret_val: bool = False
 
-        logger.info(f'Start of GOA data file processing.')
-
         # and get a reference to the data gatherer
         gd = GetData()
 
+        # get the uniprot kb ids that were curated by swiss-prot
+        swiss_prots: set = gd.get_swiss_prot_id_set(data_file_path)
+
         # get the data file
-        actual_count: int = gd.get_goa_files(data_file_path, [data_file_name], '/pub/databases/GO/goa', '/HUMAN/')
+        actual_count: int = gd.get_goa_files(data_file_path, [data_file_name], '/pub/databases/GO/goa', ftp_dir_path)
 
         # did we get all the files
         if actual_count == 1:
@@ -73,40 +75,46 @@ class GOALoader:
                 out_edge_f.write(f'id\tsubject\trelation_label\tedge_label\tobject\n')
 
                 # parse the data
-                self.parse_data_file(os.path.join(data_file_path, data_file_name), out_node_f, out_edge_f)
+                self.parse_data_file(os.path.join(data_file_path, data_file_name), out_node_f, out_edge_f, swiss_prots)
 
-                # remove the data file
-                os.remove(os.path.join(data_file_path, data_file_name))
+                # do not remove the file if in debug mode
+                if logger.level != logging.DEBUG:
+                    # remove the data file
+                    os.remove(os.path.join(data_file_path, data_file_name))
 
                 logger.debug(f'File parsing complete.')
 
                 # set the return flag
                 ret_val = True
         else:
-            logger.error(f'Error retieving file {data_file_name}.')
+            logger.error(f'Error: Retrieving file {data_file_name} failed.')
 
+        logger.info(f'GOALoader - Processing complete.')
+
+        # return the pass/fail flag to the caller
         return ret_val
 
-    def parse_data_file(self, infile_path: str, out_node_f, out_edge_f):
+    def parse_data_file(self, infile_path: str, out_node_f, out_edge_f, swiss_prots: set):
         """
         Parses the data file for graph nodes/edges and writes them out the KGX tsv files.
 
         :param infile_path: the name of the intact file to process
         :param out_edge_f: the edge file pointer
         :param out_node_f: the node file pointer
+        :param swiss_prots: the list of uniprot ids that have been swiss curated
         :return:
         """
 
         with gzip.open(infile_path, 'r') as zf:
             # read the file and make the list
-            node_list: list = self.get_node_list(zf)
+            node_list: list = self.get_node_list(zf, swiss_prots)
 
-            logger.info(f'Node list loaded with {len(node_list)} entries.')
+            logger.debug(f'Node list loaded with {len(node_list)} entries.')
 
             # de-dupe the list
             total_nodes = [dict(t) for t in {tuple(d.items()) for d in node_list}]
 
-            logger.info(f'Node list duplicates removed, now loaded with {len(total_nodes)} entries.')
+            logger.debug(f'Node list duplicates removed, now loaded with {len(total_nodes)} entries.')
 
             # normalize the group of entries on the data frame.
             nnu = NodeNormUtils()
@@ -114,7 +122,7 @@ class GOALoader:
             # normalize the node data
             total_nodes = nnu.normalize_node_data(total_nodes)
 
-            logger.info('Creating edges.')
+            logger.debug('Creating edges.')
 
             # create a data frame with the node list
             df: pd.DataFrame = pd.DataFrame(total_nodes, columns=['grp', 'node_num', 'id', 'name', 'category', 'equivalent_identifiers'])
@@ -122,13 +130,13 @@ class GOALoader:
             # get the list of unique edges
             final_edges: set = self.get_edge_set(df)
 
-            logger.info(f'{len(final_edges)} unique edges found, creating KGX edge file.')
+            logger.debug(f'{len(final_edges)} unique edges found, creating KGX edge file.')
 
             # write out the unique edges
             for item in final_edges:
                 out_edge_f.write(hashlib.md5(item.encode('utf-8')).hexdigest() + item)
 
-            logger.info(f'De-duplicating {len(total_nodes)} nodes')
+            logger.debug(f'De-duplicating {len(total_nodes)} nodes')
 
             # init a set for the node de-duplication
             final_node_set: set = set()
@@ -137,12 +145,12 @@ class GOALoader:
             for row in total_nodes:
                 final_node_set.add(f"{row['id']}\t{row['name']}\t{row['category']}\t{row['equivalent_identifiers']}\n")
 
-            logger.info(f'Creating KGX node file with {len(final_node_set)} nodes.')
+            logger.debug(f'Creating KGX node file with {len(final_node_set)} nodes.')
 
             for row in final_node_set:
                 out_node_f.write(row)
 
-        logger.info(f'GOA data parsing and KGX file creation complete.\n')
+        logger.debug(f'GOA data parsing and KGX file creation complete.\n')
 
         return True
 
@@ -171,7 +179,7 @@ class GOALoader:
 
             # if we dont get a set of 3 something is odd (but not necessarily bad)
             if len(rows) != 3:
-                logger.error(f'Mis-matched node grouping. {rows}')
+                logger.error(f'Error: Mis-matched node grouping. {rows}')
 
             # for each row in the triplet
             for row in rows.iterrows():
@@ -198,6 +206,7 @@ class GOALoader:
             relation_label: str = ''
             src_node_id: str = ''
             obj_node_id: str = ''
+            valid_type = True
 
             # find the predicate and edge relationships
             if node_3_type.find('molecular_activity') > -1:
@@ -212,9 +221,14 @@ class GOALoader:
                 relation_label = 'has_part'
                 src_node_id = node_3_id
                 obj_node_id = node_1_id
+            else:
+                valid_type = False
+                logger.error(f'Error: Unrecognized node 3 type <{node_3_type}> for {node_3_id}')
 
-            # create the KGX edge data for nodes 1 and 3
-            edge_set.add(f'\t{src_node_id}\t{relation_label}\t{relation_label}\t{obj_node_id}\n')
+            # was this a good value
+            if valid_type:
+                # create the KGX edge data for nodes 1 and 3
+                edge_set.add(f'\t{src_node_id}\t{relation_label}\t{relation_label}\t{obj_node_id}\n')
 
         logger.debug(f'{len(edge_set)} unique edges identified.')
 
@@ -222,10 +236,11 @@ class GOALoader:
         return edge_set
 
     @staticmethod
-    def get_node_list(fp) -> list:
+    def get_node_list(fp, swiss_prots: set) -> list:
         """ loads the nodes from the file handle passed
 
         :param fp: open file pointer
+        :param swiss_prots: list of swiss-prot curated uniport id entries
         :return: list of nodes for further processing
         """
 
@@ -241,34 +256,36 @@ class GOALoader:
             if line[0] == '!' or line[0][0] == '!':
                 continue
 
-            try:
-                # an example record looks like this
-                """ UniProtKB       O73942  apeI            GO:0004518      GO_REF:0000043  IEA     UniProtKB-KW:KW-0540    F       
-                    Homing endonuclease I-ApeI      apeI|APE_1929.1 protein 272557  20200229        UniProt """
+            # is this a swiss-port curated entry
+            if line[1] in swiss_prots:
+                try:
+                    # an example record looks like this
+                    """ UniProtKB       O73942  apeI            GO:0004518      GO_REF:0000043  IEA     UniProtKB-KW:KW-0540    F       
+                        Homing endonuclease I-ApeI      apeI|APE_1929.1 protein 272557  20200229        UniProt """
 
-                # create a unique group identifier
-                grp: str = f'{line[DATACOLS.DB_Object_ID.value]}{line[DATACOLS.GO_ID.value]}{line[DATACOLS.Taxon_Interacting_taxon.value]}'
+                    # create a unique group identifier
+                    grp: str = f'{line[DATACOLS.DB_Object_ID.value]}{line[DATACOLS.GO_ID.value]}{line[DATACOLS.Taxon_Interacting_taxon.value]}'
 
-                # create node type 1
-                """ A gene with identifier UniProtKB:O73942, and name "apeI", and description "Homing endonuclease I-ApeI". """
-                node_list.append({'grp': grp, 'node_num': 1, 'id': f'{line[DATACOLS.DB.value]}:{line[DATACOLS.DB_Object_ID.value]}', 'name': f'{line[DATACOLS.DB_Object_Symbol.value]}', 'category': '', 'equivalent_identifiers': f'{line[DATACOLS.DB.value]}:{line[DATACOLS.DB_Object_ID.value]}'})
+                    # create node type 1
+                    """ A gene with identifier UniProtKB:O73942, and name "apeI", and description "Homing endonuclease I-ApeI". """
+                    node_list.append({'grp': grp, 'node_num': 1, 'id': f'{line[DATACOLS.DB.value]}:{line[DATACOLS.DB_Object_ID.value]}', 'name': f'{line[DATACOLS.DB_Object_Symbol.value]}', 'category': '', 'equivalent_identifiers': f'{line[DATACOLS.DB.value]}:{line[DATACOLS.DB_Object_ID.value]}'})
 
-                # create node type 2
-                """ An organism_taxon with identifier NCBITaxon:272557. This one should node normalize fine, returning the correct names. """
-                # get the taxon id
-                taxon_id: str = line[DATACOLS.Taxon_Interacting_taxon.value]
+                    # create node type 2
+                    """ An organism_taxon with identifier NCBITaxon:272557. This one should node normalize fine, returning the correct names. """
+                    # get the taxon id
+                    taxon_id: str = line[DATACOLS.Taxon_Interacting_taxon.value]
 
-                # always take the first one in a list, if the taxon if starts with taxon remove it.
-                taxon_id = taxon_id.split('|')[0].replace('taxon:', '')
+                    # always take the first one in a list, if the taxon if starts with taxon remove it.
+                    taxon_id = taxon_id.split('|')[0].replace('taxon:', '')
 
-                # create the node
-                node_list.append({'grp': grp, 'node_num': 2, 'id': f'NCBITaxon:{taxon_id}', 'name': '', 'category': '', 'equivalent_identifiers': ''})
+                    # create the node
+                    node_list.append({'grp': grp, 'node_num': 2, 'id': f'NCBITaxon:{taxon_id}', 'name': '', 'category': '', 'equivalent_identifiers': ''})
 
-                # create node type 3
-                """ A node for the GO term GO:0004518. It should normalize, telling us the type/name. """
-                node_list.append({'grp': grp, 'node_num': 3, 'id': f'{line[DATACOLS.GO_ID.value]}', 'name': '', 'category': '', 'equivalent_identifiers': ''})
-            except Exception as e:
-                logger.error(f'Exception: {e}')
+                    # create node type 3
+                    """ A node for the GO term GO:0004518. It should normalize, telling us the type/name. """
+                    node_list.append({'grp': grp, 'node_num': 3, 'id': f'{line[DATACOLS.GO_ID.value]}', 'name': '', 'category': '', 'equivalent_identifiers': ''})
+                except Exception as e:
+                    logger.error(f'Error: Exception: {e}')
 
         # return the list to the caller
         return node_list
@@ -295,4 +312,4 @@ if __name__ == '__main__':
     vp = GOALoader()
 
     # load the data files and create KGX output
-    vp.load(data_dir, 'goa_human.gaf.gz', 'Human_GOA')
+    vp.load(data_dir, '/HUMAN/', 'goa_human.gaf.gz', 'Human_GOA')
